requirements.txt
appdirs==1.4.3
crlibm==1.0.3
cycler==0.10.0
ete3==3.0.0b35
functools32==3.2.3.post2
matplotlib==2.0.2
mpmath==0.19
numpy==1.12.1
olefile==0.44
packaging==16.8
Pillow==4.1.1
progressbar==2.3
pyinterval==1.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
scipy==0.19.0
six==1.10.0
subprocess32==3.2.7
sympy==1.0
terminaltables==3.1.0
